Change some tools in utils to backend
